Added spell rolls
Spell items have been added in, when their roll is called, it opens a dialog popup asking if the roll will use a component, and how many dice the player would like to roll. When rolling with doubles of a dice, or more, they are added up and calculate the miscast and outputs that with the role into the chat log.